requirements.txt
Flask
line-bot-sdk
ollama
python-dotenv
requests